Adding authorization checks for user actions delete and edit posts

+++ section_25_REST_APIs_Advanced/backend/controllers/feed.js
@@ -76,7 +76,7 @@ exports.createPost = (req, res, next) => {
                     _id: creator._id,
                     name: creator.name
                 }
-            });
+            }); 
         })
         .catch(err => {
             if (!err.statusCode) {
@@ -124,13 +124,17 @@ exports.updatePost = (req, res, next) => {
         err.statusCode = 422;
         throw err;
     }
-
     Post.findById(postId)
         .then(post => {
             if (!post) {
                 const err = new Error("Post does not exist!");
                 err.statusCode = 404;
                 throw err;
+            }
+            if(post.creator.toString() !== req.userId){
+                const error = new Error("Not authorised to update this post!");
+                error.statusCode = 403;
+                throw error;
             }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
@@ -160,7 +164,13 @@ exports.deletePost = (req, res, next) => {
                 err.statusCode = 404;
                 throw err;
             }
-            //check logged in user
+
+            if(post.creator.toString() !== req.userId){
+                const error = new Error("Not authorised to update this post!");
+                error.statusCode = 403;
+                throw error;
+            }
+            
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
